Setting region and timestamp_hr as primary key
The `primary_key` is only overwritten in the online feature store. The offline
feature store uses a combination of `primary_key` and `event_time`. Overwriting
by the hour should be ok since earthquake data is not expected to have multiple
events in the hour. There are exceptions to this however.

We will revisit this perhaps with a rolling window in the future.